requirements.txt
numpy==2.2.2
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.1.0
streamlit==1.41.1
torch==2.6.0
torchvision==0.21.0
